fix(ycm): fix compilation errors on Arch
There are a number of issues that were affecting YCM on Arch:

- The system libraries are too new to build YCM
- YCM assumes that /usr/bin/python is Python 2

We fix the first issue by not building with the system libraries. YCM has a
fancy caching mechanism for libclang now, so download times are no longer an
issue.

The second issue is fixed by being explicit about which interpreter we wish to
use.